Move functions to fetchData and clean up
