+++ app/dashboard.tsx
@@ -7,10 +7,8 @@ import {
   _isAboutWelcomeTemplate,
 } from "../lib/messageUtils.ts";
 
-import { NimbusRecipeCollection } from "../lib/nimbusRecipeCollection";
 import { _substituteLocalizations } from "../lib/experimentUtils.ts";
 
-import { NimbusRecipe } from "../lib/nimbusRecipe.ts";
 import { MessageTable } from "./message-table";
 
 import { MenuButton } from "@/components/ui/menubutton.tsx";
@@ -18,49 +16,8 @@ import { InfoPopover } from "@/components/ui/infopopover.tsx";
 import { Timeline } from "@/components/ui/timeline.tsx";
 import { Platform } from "@/lib/types";
 
-export const isLookerEnabled = process.env.IS_LOOKER_ENABLED === "true";
-
 const hidden_message_impression_threshold =
   process.env.HIDDEN_MESSAGE_IMPRESSION_THRESHOLD;
-
-/**
- * A sorting function to sort messages by their start dates in descending order.
- * If one or both of the recipes is missing a start date, they will be ordered
- * identically since there's not enough information to properly sort them by
- * date.
- *
- * @param a Nimbus recipe to compare with `b`.
- * @param b Nimbus recipe to compare with `a`.
- * @returns -1 if the start date for message a is after the start date for
- *          message b, zero if they're equal, and 1 otherwise.
- */
-export function compareDatesFn(a: NimbusRecipe, b: NimbusRecipe): number {
-  if (a._rawRecipe.startDate && b._rawRecipe.startDate) {
-    if (a._rawRecipe.startDate > b._rawRecipe.startDate) {
-      return -1;
-    } else if (a._rawRecipe.startDate < b._rawRecipe.startDate) {
-      return 1;
-    }
-  }
-
-  // a must be equal to b
-  return 0;
-}
-
-export async function getMsgRolloutCollection(
-  recipeCollection: NimbusRecipeCollection,
-): Promise<NimbusRecipeCollection> {
-  const msgRolloutRecipeCollection = new NimbusRecipeCollection();
-  msgRolloutRecipeCollection.recipes = recipeCollection.recipes
-    .filter((recipe) => recipe.usesMessagingFeatures() && !recipe.isExpRecipe())
-    .sort(compareDatesFn);
-  console.log(
-    "msgRolloutRecipeCollection.length = ",
-    msgRolloutRecipeCollection.recipes.length,
-  );
-
-  return msgRolloutRecipeCollection;
-}
 
 interface ReleasedTableProps {
   platform: string;
